Conflicts:
	src/game/SpellMgr.cpp

+++ src/game/Pet.cpp
@@ -3218,7 +3218,7 @@ void Pet::Regenerate(Powers power, uint32 diff)
 
     if (curValue < 0)
         curValue = 0;
-    else if (curValue > maxValue)
+    else if (curValue > int32(maxValue))
         curValue = maxValue;
 
     SetPower(power, curValue);
